#16: Support proper query timeouts
It was already possible to set a global and local maximum execution
time, however, this was not properly enforced by the query engine.

Now we keep track of the runtime of a query, and whenever blocking we
set the timeout to the maximum remaining time. If an interrupt or a
timeout occurs, this is propagated to the result iteration as
QueryInterruptedException (thus indicating a timeout).

The test infrastructure has been adjusted to simulate latency.

Note that specialized Query implementations have been introduced to
allow handing over the query timeout to the FedX engine.