#22: Prepare FedX initialization for resolving external repositories
In some use cases it might be desired to create a federation with
externally managed repositories (e.g. in a federation running in the
RDF4J workbench).

This change is a preparation for the functionality, and supports to
configure a RepositoryResolver for resolving such endpoints. Note that
documentation and concrete examples will be added in a follow up change.

As of this change FedX distinguishes between Managed and Resolvable
endpoints, where the lifecycle of the Repository instance is managed by
FedX in the first case.

Note that the FedX factory has been revised for easier configuration.

This change is partially not backwards compatible to previous releases.