#22: support to use resolvable endpoints
As of this change it is possible to configure the federation with
resolvable endpoints, that are resolved against a configured
RepositoryResolver.

In the RDF4J server for instance the repository manager can be used as a
RepositoryResolver, i.e. it is possible to reference any repository
managed in the RDF4J server.

Note that the RDF4J server's repository manager is automatically
injected into FedX, when the respective resolver bean is registered in
the RDF4J webapp.