#24: make sure that locations are relative to the base directory
This change makes sure that all locations (e.g. cache.db) are
interpreted relatively to the configured base directory.

By default the base directory is the current execution directory of the
process.

When deployed in the RDF4J Server the base directory is reconfigured to
the data directory of the federation.