#25: use fresh RepositoryConnection for interacting with members
As of this change by default we use a fresh RepositoryConnection for
each operation on a federation member.

Note that interfaces and connection handling has been adjusted, that
connections are properly closed after consumption, particularly in
erroneous cases.

It is configurable to re-activate the Singleton connection behavior
using the 'endpoint.useSingletonConnection' config setting.