#30: fix potential deadlock in OPTIONAL evaluation
In case there are large intermediate result sets in the evaluation of a
query, the left join could run into a deadlock.

This is resolved as of this change by using a controlled execution using
a separate FIFO based thread pool for left join tasks.

In addition some javadoc and leftovers are resolved.